+++ scripts/format.sh
@@ -1,4 +1,4 @@
 #! /bin/bash
 
 black vulcan tests
-
+isort --profile black vulcan tests